Revert "Feature/email intergration moment client secret validation Ranindu's 25th Tuesday commit"
