+++ api/email_filtering_and_info_generation/read_emails.py
@@ -19,11 +19,10 @@ from api.email_filtering_and_info_generation.topic_identification import identif
 from api.email_filtering_and_info_generation.suggestions_identification import identify_and_summarize_suggestions
 
 
-from api.email_authorization.services import refresh_token, is_token_valid, login_async
 
-from api.email_filtering_and_info_generation.services import get_all_reading_accounts, router
-from api.email_authorization.services import update_authorization_uri
-from api.email_filtering_and_info_generation.services import send_email_message,get_reading_emails_array
+
+from api.email_filtering_and_info_generation.routes import get_all_reading_accounts, router, update_authorization_uri
+from api.email_filtering_and_info_generation.routes import send_email_message,get_reading_emails_array
 
 from fastapi import FastAPI, Request
 from fastapi.responses import RedirectResponse
@@ -52,7 +51,103 @@ last_email_read_time = ""
 
 # input the email_acc_array and then return the updated email_acc_array
 # email_acc_array=integrateEmail()
+def refresh_token(token_path):
+    # Load the token data from the JSON file
+    with open(token_path, 'r') as token_file:
+        token_data = json.load(token_file)
 
+    # Refresh token request parameters
+    refresh_token = token_data['refresh_token']
+    client_id = token_data['client_id']
+    client_secret = token_data['client_secret']
+    token_uri = token_data['token_uri']
+
+    # Define the request payload
+    payload = {
+        'client_id': client_id,
+        'client_secret': client_secret,
+        'refresh_token': refresh_token,
+        'grant_type': 'refresh_token',
+    }
+
+    # Make the request to refresh the token
+    response = requests.post(token_uri, data=payload)
+
+    if response.status_code == 200:
+        new_token_data = response.json()
+        # Update the token data with the new token information
+        token_data['token'] = new_token_data['access_token']
+        # Update the expiry time (typically expires_in is in seconds)
+        expiry_time = datetime.utcnow() + timedelta(seconds=new_token_data['expires_in'])
+        token_data['expiry'] = expiry_time.isoformat() + 'Z'  # Adding 'Z' to indicate UTC
+
+        # Save the updated token data back to the JSON file
+        with open(token_path, 'w') as token_file:
+            json.dump(token_data, token_file, indent=4)
+
+        print("Token refreshed successfully!")
+    else:
+        print(f"Failed to refresh token: {response.status_code}")
+        print(response.json())
+
+
+
+def is_token_valid(token_file):
+    try:
+        with open(token_file, 'r') as file:
+            token_data = json.load(file)
+        
+        expiration_timestamp = token_data['expiry'].rstrip('Z')
+        expiration_datetime = datetime.fromisoformat(expiration_timestamp).replace(tzinfo=timezone.utc)
+        current_datetime = datetime.now(timezone.utc)
+
+        if current_datetime < expiration_datetime - timedelta(minutes=3):
+            print("Token is still valid.")
+            return True
+        else:
+            print("Token has expired.")
+            if os.path.exists(token_file):
+                # Delete the token file
+                os.remove(token_file)    
+            return False
+    except FileNotFoundError:
+        print("Token file not found.")
+        return False
+    except KeyError:
+        print("Invalid token file format.")
+        return False
+
+# Initialize the OAuth flow
+def init_oauth_flow(client_secrets_file: str, redirect_uri: str):
+    flow = Flow.from_client_secrets_file(
+        client_secrets_file,
+        scopes = [
+            'https://www.googleapis.com/auth/gmail.modify',
+            'https://www.googleapis.com/auth/gmail.settings.basic'
+        ],
+
+        redirect_uri=redirect_uri
+    )
+    return flow
+
+
+async def login_async(id: int, email_acc_address:str):
+    os.environ['OAUTHLIB_INSECURE_TRANSPORT'] = '1'
+
+    client_secrets_file = f'api/email_filtering_and_info_generation/credentialsForEmails/credentialsForEmail{id}/client_secret.json'
+    redirect_uri = f'http://127.0.0.1:8000/email/info_and_retrieval/callback?id={id}'
+    flow = init_oauth_flow(client_secrets_file, redirect_uri)
+
+    state = os.urandom(16).hex()  # Generate a random state
+    state_store[state] = id  # Store the state with the associated id
+
+    authorization_url, _ = flow.authorization_url(
+        access_type='offline',
+        include_granted_scopes='true',
+        state=state
+    )
+    print("authorization_url: ",authorization_url)
+    await update_authorization_uri(authorization_url, email_acc_address)
 
 
 
@@ -345,51 +440,25 @@ async def repeat_every_10mins():
     while True:
         time.sleep(max(0, next_time - time.time()))
         try:
+            print("read_all_new_emails happening")
             new_email_msg_array = []
-            last_email_read_time = ""
-            await read_all_new_emails(new_email_msg_array, last_email_read_time)
-            print("\n")
-            print("\n")
-            # print("RAW READ NEW EMAIL MESG ARRAY", new_email_msg_array)
+            await read_all_new_emails(new_email_msg_array)
+            #print(new_email_msg_array)
             mask_email_messages(new_email_msg_array)
-            #print("MASKED EMAIL MESSAGES", new_email_msg_array)
-            # await format_email_bodies(new_email_msg_array)
-            # print("\n")
-            # print("\n formatted email bodies")
-            # print("\n")
-            # print("formatted EMAIL BODIES----------",new_email_msg_array)
-            # identify_sentiments(new_email_msg_array)
-            # identify_criticality(new_email_msg_array)
-            # print("\n")
-            # print("\n criticality identified emails")
-            # print("\n")
+            format_email_bodies(new_email_msg_array)
+            identify_sentiments(new_email_msg_array)
+            identify_criticality(new_email_msg_array)
+            #print(new_email_msg_array)
+            await identify_notifcations(new_email_msg_array)
             # print(new_email_msg_array)
-            # # #await identify_notifcations(new_email_msg_array)
-            # await identify_topics(new_email_msg_array)
-            # print("\n")
-            # print("\n Products identified emails")
-            # print("\n")
-            # print("emails with topics",new_email_msg_array)
-            # # # print(new_email_msg_array)
-            # # #print(new_email_msg_array)
-            # await identify_issues_inquiries_and_checking_status(new_email_msg_array) # make sure to perform this before the coversations_summarizing
-            # # print("---------------------------------------finished identify issues------------------------------")
-            # print("\n")
-            # print("\n Issues and Inquiries identified emails")
-            # print("\n")
-            # print(new_email_msg_array)
-            # await identify_and_summarize_suggestions(new_email_msg_array)
-            # print("\n")
-            # print("\n Suggestions identified emails")
-            # print("\n")
-            # print(new_email_msg_array)
-            # # print("---------------------------------------finished identifying summaries------------------------")
-            # # print(new_email_msg_array)
-            # await push_new_emails_to_DB(new_email_msg_array)
-            # print("database update successful")
-            summarize_conversations(new_email_msg_array)
-            print("conversation summaries identified")
-                    
+            identify_topics(new_email_msg_array)
+            print(new_email_msg_array)
+            await push_new_emails_to_DB(new_email_msg_array)
+            
+            await identify_issues_inquiries_and_checking_status(new_email_msg_array) # make sure to perform this before the coversations_summarizing
+            await summarize_conversations(new_email_msg_array)
+            await identify_and_summarize_suggestions(new_email_msg_array)
+            
         except Exception as e:
             print("Error:", e)
         # Calculate the next execution time
