+++ api/email_filtering_and_info_generation/routes.py
@@ -1,14 +1,13 @@
 from datetime import datetime, timedelta
 import json
 import os
-from typing import List
 from fastapi import APIRouter, HTTPException
 from pymongo import MongoClient
 # import sys
 # sys.path.append('..') 
 from api.email_filtering_and_info_generation.configurations.database import collection_email_msgs
 from api.email_filtering_and_info_generation.configurations.database import collection_triger_events,collection_trigers,collection_readingEmailAccounts,collection_suggestions, collection_conversations, collection_issues, collection_inquiries, collection_overdue_trigger_events, collection_configurations
-from api.email_filtering_and_info_generation.models import Convo_summary, Email_msg, InquiryInDB, IssueInDB, Overdue_trig_event,Trigger_event,Reading_email_acc,Suggestion
+from api.email_filtering_and_info_generation.models import Convo_summary, Email_msg, Inquiry, Issue, Overdue_trig_event,Trigger_event,Reading_email_acc,Suggestion
 from api.email_filtering_and_info_generation.schemas import individual_email_msg_serial,list_email_msg_serial
 from api.email_filtering_and_info_generation.schemas import individual_trigger_serial,list_trigger_serial
 from api.email_filtering_and_info_generation.schemas import individual_readingEmailAcc_serial, list_readingEmailAcc_serial
@@ -25,19 +24,95 @@ from pathlib import Path
 router = APIRouter()
 
 async def get_all_reading_accounts():
-    documents = collection_readingEmailAccounts.find({}, {"_id": 0, "address": 1})
+    documents = await collection_readingEmailAccounts.find({}, {"_id": 0, "address": 1})
     
     all_reading_email_accs = [doc["address"] for doc in documents]
     
     return all_reading_email_accs
 
+async def update_authorization_uri(authorization_url: str, email_acc_address: str):
+    
+    doc = {"id":2, "needToAuthorize":True, "needToAuthorizeAddress":email_acc_address, "authorization_url":authorization_url}
+    
+    existing_document = collection_configurations.find_one({"id": 2})
+    
+    if existing_document:
+         # Update the document with the new values
+         collection_configurations.update_one({"id": 2}, {"$set": doc})   
+    else: 
+    
+        collection_configurations.insert_one(doc)
+        
+        
+# Initialize the OAuth flow
+def init_oauth_flow(client_secrets_file: str, redirect_uri: str):
+    flow = Flow.from_client_secrets_file(
+        client_secrets_file,
+        scopes = [
+            'https://www.googleapis.com/auth/gmail.modify',
+            'https://www.googleapis.com/auth/gmail.settings.basic'
+        ],
+
+        redirect_uri=redirect_uri
+    )
+    return flow
+
+@router.get("/info_and_retrieval/callback")
+async def callback(request: Request, id: int):
+    
+    
+    print("Received callback request:")
+    print(f"Query parameters: {request.query_params}")
+    print(f"Path parameters: {request.path_params}")
+    print(f"Headers: {request.headers}")
+    print(f"Body: {await request.body()}")
+    print("in the CALLBACK!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!'")
+
+    try:
+        # Print the query parameters received in the request
+        print("Query Parameters:", dict(request.query_params))
+        id = 1
+        
+        # Extract query parameters
+        state = request.query_params.get('state')
+        code = request.query_params.get('code')
 
         
+        # Handle the rest of the callback logic here
+        client_secrets_file = f"api/email_filtering_and_info_generation/credentialsForEmails/credentialsForEmail{id}/client_secret.json"  
+        redirect_uri = f'http://127.0.0.1:8000/email/info_and_retrieval/callback?id={id}'
+        flow = init_oauth_flow(client_secrets_file, redirect_uri)
+        flow.fetch_token(authorization_response=str(request.url))
+        
+        credentials = flow.credentials
+        token_path = f'api/email_filtering_and_info_generation/credentialsForEmails/credentialsForEmail{id}/gmail_token.json'
+        
+        # Delete the existing file if it exists
+        if os.path.exists(token_path):
+            os.remove(token_path)
+        
+        print("credentials", credentials)
+        credentials_dict_string = credentials.to_json()
 
-
-
+        
+        # Write the new credentials
+        with open(token_path, 'w') as token_file:
+            token_file.write(credentials_dict_string)
+        
+        doc = {"id":2, "needToAuthorize":False, "needToAuthorizeAddress":"" ,"authorization_url":""}
+        collection_configurations.update_one({"id": 2}, {"$set": doc}) 
+            
+        # Optionally, you can redirect or return a response here
+        return {"message": "Authentication successful"}
+        
+    except Exception as e:
+        # Print any exceptions that occur during callback processing
+        print("Error in callback:", e)
+        raise HTTPException(status_code=500, detail="Internal server error")
+    
   
 # to send an email msg to EmailMessages collection
+@router.post("/info_and_retrieval/send_email")
 async def send_email_message(email_msg_dict: Email_msg):
     try:
         # Insert the email message dictionary into the MongoDB collection
@@ -51,6 +126,7 @@ async def send_email_message(email_msg_dict: Email_msg):
     
     
 # to send a suggestion to Suggestions collection
+@router.post("/info_and_retrieval/send_suggestion")
 async def send_suggestion(suggestion: Suggestion):
     try:
         # Insert the suggestion dictionary into the MongoDB collection
@@ -63,6 +139,7 @@ async def send_suggestion(suggestion: Suggestion):
     
 
 # to send a conversation summary to Conversations collection
+@router.post("/info_and_retrieval/send_convo_summary")
 async def send_convo_summary(convo_summary: Convo_summary):
     try:
         # Insert the conversation summary dictionary into the MongoDB collection
@@ -73,11 +150,11 @@ async def send_convo_summary(convo_summary: Convo_summary):
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-
-async def update_summary(thread_id: str, new_summary: str, new_updated_times:datetime):
+@router.put("/info_and_retrieval/update-summary/")
+async def update_summary(thread_id: str, new_summary: str):
     result =  collection_conversations.update_one(
         {"thread_id": thread_id},
-        {"$set": {"summary": new_summary, "updated_times": new_updated_times}}
+        {"$set": {"summary": new_summary}}
     )
     
     if result.matched_count == 0:
@@ -86,10 +163,11 @@ async def update_summary(thread_id: str, new_summary: str, new_updated_times:dat
     return {"message": "Summary updated successfully"}
     
 # to send an issue to Issues collection
-async def send_issue(issue: IssueInDB):
+@router.post("/info_and_retrieval/send_issue")
+async def send_issue(issue: Issue):
     try:
         # Insert the conversation summary dictionary into the MongoDB collection
-        result = collection_issues.insert_one(issue.dict())
+        result = collection_issues.insert_one(issue)
         
         # Return the ID of the inserted document
         return {"message": "Issue sent successfully", "inserted_id": str(result.inserted_id)}
@@ -97,41 +175,39 @@ async def send_issue(issue: IssueInDB):
         raise HTTPException(status_code=500, detail=str(e)) 
 
 
-
-async def update_issue_status(thread_id_to_update:str, issue_convo_summary_arr:List[dict], updated_time:datetime, new_status:str, new_ongoing_status:str, sentiment_score: float, end_time_value:datetime, new_effectiveness:str, new_efficiency:str):
+@router.post("/info_and_retrieval/update_issue_status")
+async def update_issue_status(thread_id_to_update:str, new_issue_convo_summary:str, new_status:str, end_time_value:datetime, new_effectiveness:str, new_efficiency:str):
     try:
         # Update the document with the specified thread_id
         result = collection_issues.update_one(
             {"thread_id": thread_id_to_update},
-            {"$set": {"issue_convo_summary_arr":issue_convo_summary_arr,"updated_time":updated_time,"status": new_status, 
-                      "ongoing_status":new_ongoing_status, "sentiment_score":sentiment_score,
-                      "end_time": end_time_value, "effectiveness":new_effectiveness, "efficiency":new_efficiency}}
+            {"$set": {"issue_convo_summary":new_issue_convo_summary,"status": new_status, "end_time": end_time_value, "effectiveness":new_effectiveness, "efficiency":new_efficiency}}
         )
+        
         
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
     
     
 # to send an inquiry to Inquiries collection
-async def send_inquiry(inquiry: InquiryInDB):
+@router.post("/info_and_retrieval/send_inquiry")
+async def send_inquiry(inquiry: Inquiry):
     try:
         # Insert the conversation summary dictionary into the MongoDB collection
-        result = collection_inquiries.insert_one(inquiry.dict())
+        result = collection_inquiries.insert_one(inquiry)
         
         # Return the ID of the inserted document
         return {"message": "Inquiry sent successfully", "inserted_id": str(result.inserted_id)}
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))    
 
-
-async def update_inquiry_status(thread_id_to_update:str, new_inquiry_convo_summary:str, updated_time:datetime, new_status:str, new_ongoing_status:str, sentiment_score: float,end_time_value:datetime, new_effectiveness:str, new_efficiency:str):
+@router.post("/info_and_retrieval/update_inquiry_status")
+async def update_inquiry_status(thread_id_to_update:str, new_inquiry_convo_summary:str, new_status:str, end_time_value:datetime, new_effectiveness:str, new_efficiency:str):
     try:
         # Update the document with the specified thread_id
         result = collection_inquiries.update_one(
             {"thread_id": thread_id_to_update},
-            {"$set": {"inquiry_convo_summary":new_inquiry_convo_summary, "updated_time":updated_time, 
-                      "status": new_status, "ongoing_status":new_ongoing_status, "sentiment_score":sentiment_score,
-                      "end_time": end_time_value, "effectiveness":new_effectiveness, "efficiency":new_efficiency}}
+            {"$set": {"inquiry_convo_summary":new_inquiry_convo_summary, "status": new_status, "end_time": end_time_value, "effectiveness":new_effectiveness, "efficiency":new_efficiency}}
         )
         
         
@@ -140,6 +216,7 @@ async def update_inquiry_status(thread_id_to_update:str, new_inquiry_convo_summa
     
     
 # to send a trigger event to TriggerEvents collection
+@router.post("/info_and_retrieval/send_trigger_event")
 async def send_trig_event(trig_event_dict: Trigger_event):
     try:
         # Insert the tig_event dictionary into the MongoDB collection
@@ -151,6 +228,9 @@ async def send_trig_event(trig_event_dict: Trigger_event):
         raise HTTPException(status_code=500, detail=str(e))
     
 
+
+
+@router.post("/")
 async def post_email_msg(email_msg: Email_msg):
     collection_email_msgs.insert_one(email_msg.dict())
     
@@ -158,18 +238,22 @@ async def post_email_msg(email_msg: Email_msg):
 
 # get all the triggers from the TriggerEvents collection
 
+@router.get("/info_and_retrieval/get_triggers")
 async def get_triggers_array():
     triggers_array = list_trigger_serial(collection_trigers.find())
     return triggers_array
 
 
 # get all the triggers from the readingEmailAccounts collection
+
+@router.get("/info_and_retrieval/get_reading_email_accounts")
 async def get_reading_emails_array():
     email_acc_array = list_readingEmailAcc_serial(collection_readingEmailAccounts.find())
     return email_acc_array
     
 
 # send a new email into readingEmailAccounts collection
+@router.post("/info_and_retrieval/send_reading_email_account")
 async def send_reading_email_account(readingEmailAcc: Reading_email_acc):
     try:
         # Insert the tig_event dictionary into the MongoDB collection
@@ -246,32 +330,6 @@ async def get_overdue_issues(overdue_margin_time):
     result = list(collection_issues.aggregate(pipeline))
     return result
 
-async def get_overdue_inquiries(overdue_margin_time):
-    # Calculate the threshold date (start_time + 14 days)
-    threshold_date = datetime.utcnow() - timedelta(days=overdue_margin_time)
-
-    # Aggregation pipeline to filter documents with status 'ongoing' and start_time older than 14 days
-    pipeline = [
-        {
-            "$match": {
-                "status": "ongoing",
-                "start_time": {"$lt": threshold_date},
-                "isOverdue":False
-            }
-        },
-        {
-            "$project": {
-                "_id": 0,
-                "thread_id": 1,
-                "recepient_email":1
-            }
-        }
-    ]
-
-    # Execute the aggregation pipeline
-    result = list(collection_inquiries.aggregate(pipeline))
-    return result
-
 
 async def send_overdue_trigger_event(overdue_trigger_event: Overdue_trig_event):
     try:
@@ -302,5 +360,37 @@ async def getProductsList():
 
 
 
+# ----------------------------------------  email permissions authorization API calls -----------------------------------------
 
+
+@router.get("/info_and_retrieval/get_need_for_authorization")
+async def get_need_for_authorization():
+    
+    document = collection_configurations.find_one({"id": 2})
+
+    if document:
+        need_to_authorize = document.get("needToAuthorize")
+        if need_to_authorize is not None:
+            result =  {"needToAuthorize": need_to_authorize}
+            return JSONResponse(content=result)
+        else:
+            raise HTTPException(status_code=404, detail="Field 'needToAuthorize' not found in document")
+    else:
+        raise HTTPException(status_code=404, detail="Document with id 2 not found")
+    
+
+@router.get("/info_and_retrieval/get_authorization_info")
+async def get_authorization_info():
+    
+    document = collection_configurations.find_one({"id": 2})
+
+    if document:
+        need_to_authorize = document.get("needToAuthorize")
+        needToAuthorizeAddress = document.get("needToAuthorizeAddress")
+        authorization_url = document.get("authorization_url")
+        
+        result = {"need_to_authorize":need_to_authorize, "needToAuthorizeAddress":needToAuthorizeAddress, "authorization_url":authorization_url}
+        return JSONResponse(content=result)
+    else:
+        raise HTTPException(status_code=404, detail="Document with id 2 not found")
