+++ main.py
@@ -10,10 +10,9 @@ import time
 import uvicorn  # debugging
 
 from fastapi import FastAPI
-from api.email_filtering_and_info_generation.check_notifications import check_notifications_for_managers
 from api.summary.routes import router as conversation_router
 
-from api.email_authorization.routes import router as authorization_router
+from api.email_filtering_and_info_generation.routes import router as retrieval_and_info_router
 from api.filtering.routes import router as filtering_router
 from api.settings.routes import router as settings_router
 from api.dashboard.routes import router as dashboard_router
@@ -21,8 +20,6 @@ from api.suggestions_page.routes import router as suggestions_router
 from api.email_filtering_and_info_generation.read_emails import repeat_every_10mins
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import RedirectResponse
-
-from utils.sliding_window_deletion import slide_the_time_window
 
 app = FastAPI()
 
@@ -40,35 +37,25 @@ EMAIL_PREFIX = "/email"
 app.include_router(conversation_router, prefix=EMAIL_PREFIX)
 app.include_router(settings_router, prefix=EMAIL_PREFIX)
 app.include_router(filtering_router, prefix=EMAIL_PREFIX)
-app.include_router(authorization_router, prefix=EMAIL_PREFIX)
+app.include_router(retrieval_and_info_router, prefix=EMAIL_PREFIX)
 app.include_router(dashboard_router,prefix=EMAIL_PREFIX)
 app.include_router(suggestions_router,prefix=EMAIL_PREFIX)
 
 
-def retrieving_emails_loop():
+def run_in_thread():
     asyncio.run(repeat_every_10mins())
 
 def check_notifications_loop():
-    asyncio.run(check_notifications_for_managers())
- 
- 
-def slide_time_window_loop():
-    try:
-        asyncio.run(slide_the_time_window())
-    except Exception as e:
-        print(f"An error occurred: {e}")
+    None
 
 # @app.on_event("startup")
 # async def on_startup():
     
 #     # start the continous loop to extract emails in a new thread
-#     threading.Thread(target=retrieving_emails_loop, args=(), daemon=True).start()
+#     threading.Thread(target=run_in_thread, args=(), daemon=True).start()
 
 #     # start the continous loop to check notifications in a new thread
 #     #threading.Thread(target=check_notifications_loop, args=(), daemon=True).start()
-
-    # start the continous loop to delete data and slide the time window in a new thread
-    # threading.Thread(target=slide_time_window_loop, args=(), daemon=True).start()
